RandomAccessor

+++ include/UECS/World.h
@@ -30,7 +30,7 @@ namespace Ubpa::UECS {
 		// 4. run commands in main thread
 		void Update();
 
-		void AddCommand(std::function<void()> command, size_t layer = 0);
+		void AddCommand(std::function<void()> command, int layer = 0);
 
 		// after running Update()
 		// you can use graphviz to vistualize the graph
@@ -120,7 +120,7 @@ namespace Ubpa::UECS {
 		Pool<Job> jobPool;
 
 		// command
-		std::map<size_t, std::vector<std::function<void()>>> commandBuffer;
+		std::map<int, std::vector<std::function<void()>>> commandBuffer;
 		std::mutex commandBufferMutex;
 		void RunCommands();
